Extend `.bashrc` via host `.bashrc_extra` file
It is known that builtin .bashrc file shipped within the image hooks the
.bash_aliases file, if present. We can bind mount .bashrc_extra directly
to the .bash_aliases file, but an undefined behavior is expected, since
Docker doesn't care about i-node change when modifying directly a bind
mounted file from the host, thus the file won't update in the container.
To workaround it, we shall mount it via a directory.

Signed-off-by: iusmac <iusico.maxim@libero.it>