run: prevent interaction with Docker container linked to other project
This patch ensures that interactions with the Docker container is
restricted to the projected that launched it. Previously, actions such
as "Jump inside" would attach to an already-running container without
warning the user.
Another example, the "Build ROM" action will start/resume building the
ROM for the project that launched the container.

These behaviors are totally misleading and lead to unwanted results.

Signed-off-by: iusmac <iusico.maxim@libero.it>