run: handle silent exit during argument processing
Previously, launching with the following command will cause the script
to silently exit due to unrecognized argument:

    ./run.sh --ccache-size '40G' --invalid-argument

which is the expected behavior, but, we also expect it to print:

    Unrecognized argument: --invalid-argument

This occurs because the remaining arguments are 1, but we do 'shift 2'
which silently exists with error code 1 and causes the script to
terminate immediately due to the '-o errexit' shell option.

We don't need any further checks, just suppress the error exit when
shifting positional arguments for this particular case.

Signed-off-by: iusmac <iusico.maxim@libero.it>