run: fix self-update logic
Using git pull --recursive-submodules from the project directory won't
update the builder submodule to the latest version, but instead it will
sync the sources with the hash commit, so it still requires manually to
pull and commit the new submodule hash.
In order automate this, we'll enter into the builder directory and
explicitly pull the changes.

Signed-off-by: iusmac <iusico.maxim@libero.it>