run: fix output from `docker build/pull` sometimes being invisible
Some commands (e.g., Sources > Selective Sync) capture the output when
performing container queries, but if the docker image doesn't yet, we'll
first build it (w/ 'docker build' command) or pull it from the
repository (w/ 'docker pull' command). The issue is that the output of
building or downloading from these commands won't be seen by the user as
it's captured by some command substitution ( $(...) ).

To fix that, we'll redirect any output from docker commands to stderr.

Signed-off-by: iusmac <iusico.maxim@libero.it>