run: workaround setting Docker container host timezone (`/etc/timezone`)
We cannot rely on the '/etc/timezone' mount since it could be missing or
be a directory instead of a file. The following error is thrown during
container start when /etc/timezone is a directory:

    Error response from daemon: failed to create shim task: OCI runtime create failed: runc create failed: unable to start container process: error during container init: error mounting "/etc/timezone" to rootfs at "/etc/timezone": mount /etc/timezone:/etc/timezone (via /proc/self/fd/6), flags: 0x5001: not a directory: unknown: Are you trying to mount a directory onto a file (or vice-versa)? Check if the specified host path exists and is the expected type

We'll use $TZ environment variable when creating the container as a
workaround.

While we're at it, the $TZ detection logic has been improved: the
previous fallback to 'http://ip-api.com/line?fields=timezone' is no
longer reliable (the site is dead), so now we'll attempt to read the
'/etc/timezone' (if available) as an alternative fallback.

Signed-off-by: iusmac <iusico.maxim@libero.it>